Port `EffectTests` to the `AsyncTest` framework
